Add an optional server on port 3051 to receive patterns
This adds a server that listens for POST on port 3051 by default. It
accepts Live 1.05 and plain text formatted pattern files (the same as
the -pattern cmdline) and it adds them to the currently running world.

eg. run `sdl2-life -empty -server` and then:

curl --data-binary @./examples/glider-gun-1.05.life http://127.0.0.1:3051/

to add the glider gun example.