Change how the world size is calculated
It used to accept a window size, and a world size in rows/columns and
then would round the cell size down to what fit the smallest side. But
this could result in unused window space.

This changes it to accept the window size, in pixels, and the square
cell size in pixels. It then calculates the most cells that can fit in
the remaining space after taking the status text area into
consideration.

This also adds initial work for the -rotate and -status-top flags.